moved pbs-related functions to pbs.py
